Start work on example implementation and test suite.
Not sure the docs are right at the moment - the last test in this commit seems
to have the proposer proposing two different values for the same proposal. Need
to work out if that's because it's receiving dodgy data or if, in fact, that's
another requirement that should go in the spec.